refactor: move theme a bit more
