+++ packages/core/src/index.ts
@@ -1,4 +1,4 @@
 export * from './components';
 export * from './helpers';
-export * from './theming';
+export * from './theme';
 export * from './utils';
