+++ packages/core/src/theme/index.scss
@@ -1,0 +1,4 @@
+@forward './theme';
+@forward './themes/day';
+@forward './themes/night';
+@forward './tokens';
