+++ packages/core/src/theme/switchTheme/switchTheme.spec.ts
