--- packages/core/src/themes/index.scss
@@ -1,2 +1,0 @@
-@forward './day';
-@forward './night';
