--- packages/core/src/theming/index.scss
@@ -1,2 +1,0 @@
-@forward './theme';
-@forward './tokens';
